+++ OBC/src/obc/ConvolutionalLayer.cpp
@@ -3,20 +3,20 @@
 namespace obc {
 
 	void CrossCorrelate(
-		const std::vector<double>& A, size_t a_offset, size_t a_height, size_t a_width,
-		const std::vector<double>& B, size_t b_offset, size_t b_height, size_t b_width,
-		std::vector<double>& C, size_t c_offset, size_t c_height, size_t c_width) {
+		const std::vector<double>& A, int a_offset, int a_height, int a_width,
+		const std::vector<double>& B, int b_offset, int b_height, int b_width,
+		std::vector<double>& C, int c_offset, int c_height, int c_width) {
 
-		for (size_t i = 0; i <= a_width - b_width; i++) {
-			for (size_t j = 0; j <= a_height - b_height; j++) {
+		for (int i = 0; i <= a_width - b_width; i++) {
+			for (int j = 0; j <= a_height - b_height; j++) {
 
-				size_t c_index = c_offset + (j * c_width + i);
+				int c_index = c_offset + (j * c_width + i);
 
-				for (size_t x = 0; x < b_width; x++) {
-					for (size_t y = 0; y < b_height; y++) {
+				for (int x = 0; x < b_width; x++) {
+					for (int y = 0; y < b_height; y++) {
 
-						size_t a_index = a_offset + ((j + y) * a_width + (i + x));
-						size_t b_index = b_offset + (y * b_width + x);
+						int a_index = a_offset + ((j + y) * a_width + (i + x));
+						int b_index = b_offset + (y * b_width + x);
 
 						C[c_index] += A[a_index] * B[b_index];
 					}
@@ -64,11 +64,11 @@ namespace obc {
 
 		std::copy(biases_.begin(), biases_.end(), output_.begin());
 
-		for (size_t s = 0; s < kernel_sets_; s++) {
-			for (size_t d = 0; d < kernel_depth_; d++) {
-				size_t kernel_offset = GetKernelOffset(s, d);
-				size_t output_offset = GetOutputOffset(d);
-				size_t input_offset = d * input_width_ * input_height_;
+		for (int s = 0; s < kernel_sets_; s++) {
+			for (int d = 0; d < kernel_depth_; d++) {
+				int kernel_offset = GetKernelOffset(s, d);
+				int output_offset = GetOutputOffset(d);
+				int input_offset = d * input_width_ * input_height_;
 
 				CrossCorrelate(*input_, input_offset, input_height_, input_width_,
 								kernels_, kernel_offset, kernel_size_, kernel_size_,
